Semicolon: Style footer widgets.
For the time being, I've made these widgets center-aligned, although I'm not sure that's the best choice given the strong left-alignment of the overall design. This just feels like the most balanced way to display widgets here, particularly on large screens. Might be worth revisiting & tweaking later in the process.