Semicolon: Remove filtering functions for post frontpage.
These functions would select the two most recent featured posts, and move them to the top of the post hierarchy. I think this might be unexpected behaviour here. Since this can be acheived using sticky posts instead, I'd rather leave it to the user to decide how posts should be ordered. This way, the featured post functionality is just about assigning some prominence and visual interest to the grid.