Semicolon: Remove placeholder images for posts without featured images, and use CSS styles instead.
This sets us up to use some custom styling for different post formats, which will add interest to the page if a user hasn't uploaded many featured images.
It's also a bit more flexible, since we can change the appearance in CSS rather than by manually adjusting a static image.
We're using a CSS hack to get precise sizing on the blocks, so they don't feel awkward or out-of-place in the overall display.

I've also tweaked a bit of the spacing of images within the grid so they fit together a bit better.